LayoutTests/media/modern-media-controls/airplay-support/airplay-support.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;" autoplay></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>AirPlaySupport</code> behavior.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const container = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(container, media, null);

window.internals.setMockMediaPlaybackTargetPickerEnabled(false);

media.addEventListener("play", () => {
    media.addEventListener("webkitplaybacktargetavailabilitychanged", playbackTargetAvailabilityChanged);
    media.addEventListener("webkitcurrentplaybacktargetiswirelesschanged", currentPlaybackTargetIsWirelessChanged);

    debug("Making AirPlay routes available.");
    window.internals.setMockMediaPlaybackTargetPickerEnabled(true);
});

let playbackTargetBecameAvailable = false;
function playbackTargetAvailabilityChanged(event)
{
    if (event.availability == "available" && !playbackTargetBecameAvailable) {
        playbackTargetBecameAvailable = true;
        shouldBeTrue("mediaController.controls.airplayButton.enabled");
        shouldBeFalse("mediaController.controls.airplayButton.on");

        debug("");
        debug("Entering AirPlay playback mode.");
        window.internals.setMockMediaPlaybackTargetPickerState("Sleepy TV", "DeviceAvailable");
        media.webkitShowPlaybackTargetPicker();
    }
}

function currentPlaybackTargetIsWirelessChanged(event)
{
    if (!media.webkitCurrentPlaybackTargetIsWireless)
        return;

    shouldBeTrue("mediaController.controls.airplayButton.enabled");
    shouldBeTrue("mediaController.controls.airplayButton.on");

    debug("");
    container.remove();
    media.remove();
    finishJSTest();
}

setTimeout(finishJSTest, 3000);

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
